[feat] created fastem_grid file to store grid specific class and contants which can be used by other modules
Such a stuture prevents circular imports error for acquisition and tabs modules.
