+++ src/odemis/gui/cont/tabs/fastem_project_roas_tab.py
@@ -20,7 +20,6 @@ Odemis. If not, see http://www.gnu.org/licenses/.
 
 """
 import math
-from enum import Enum
 from typing import List, Optional
 
 import wx
@@ -34,6 +33,7 @@ from wx.grid import (
 import odemis.gui.model as guimod
 from odemis import model
 from odemis.gui.comp.fastem_roa import FastEMROA
+from odemis.gui.cont.fastem_grid import ROAColumnNames, SectionColumnNames
 from odemis.gui.cont.fastem_grid_base import (
     DEFAULT_PARENT,
     EVT_GRID_ROW_CHANGED,
@@ -42,23 +42,10 @@ from odemis.gui.cont.fastem_grid_base import (
     GridBase,
     Row,
 )
-from odemis.gui.cont.tabs.fastem_project_sections_tab import SectionColumnNames
 from odemis.gui.cont.tabs.tab import Tab
 from odemis.gui.util import call_in_wx_main
 from odemis.util import units
 from odemis.util.filename import make_compliant_string
-
-
-class ROAColumnNames(Enum):
-    NAME = "Name"
-    SLICE_IDX = "Slice Index"
-    POSX = "Position.X [m]"
-    POSY = "Position.Y [m]"
-    SIZEX = "Size.X"
-    SIZEY = "Size.Y"
-    ROT = "Rotation [°]"
-    PARENT = "Parent"
-    FIELDS = "Show Fields"
 
 
 class ROARow(Row):
